Revert MainActivity to original version and relocate test file
