+++ app/src/main/java/com/android/sample/MainActivity.kt
@@ -1,52 +1,43 @@
 package com.android.sample
 
-import android.content.pm.PackageManager
 import android.os.Bundle
-import android.util.Log
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.semantics.semantics
+import androidx.compose.ui.semantics.testTag
 import androidx.compose.ui.tooling.preview.Preview
-import com.android.sample.ui.MainScreen
+import androidx.lifecycle.viewmodel.compose.viewModel
+import com.android.sample.model.map.MapViewModel
+import com.android.sample.resources.C
+import com.android.sample.ui.map.MapScreen
 import com.android.sample.ui.theme.SampleAppTheme
-
-private const val LOCATION_PERMISSION_REQUEST_CODE = 1001
 
 class MainActivity : ComponentActivity() {
   override fun onCreate(savedInstanceState: Bundle?) {
     super.onCreate(savedInstanceState)
-    setContent { MainScreen() }
-  }
 
-  override fun onRequestPermissionsResult(
-      requestCode: Int,
-      permissions: Array<String>,
-      grantResults: IntArray
-  ) {
-    super.onRequestPermissionsResult(requestCode, permissions, grantResults)
-    if (requestCode == LOCATION_PERMISSION_REQUEST_CODE) {
-      if ((grantResults.isNotEmpty() && grantResults[0] == PackageManager.PERMISSION_GRANTED)) {
-        // Permission is granted, trigger the location fetching logic again
-        Log.d("MainActivity", "Location permission granted, fetching location")
-        fetchLocationAfterPermissionGranted()
-      } else {
-        Log.d("MainActivity", "Location permission denied")
-      }
-    }
-  }
-
-  private fun fetchLocationAfterPermissionGranted() {
-    // This will re-trigger location fetching after permission is granted
     setContent {
-      MainScreen() // Trigger composable re-composition after permission grant
+      SampleAppTheme {
+        // A surface container using the 'background' color from the theme
+        Surface(
+            modifier = Modifier.fillMaxSize().semantics { testTag = C.Tag.main_screen_container },
+            color = MaterialTheme.colorScheme.background) {
+              Greeting("Android")
+            }
+      }
     }
   }
 }
 
 @Composable
 fun Greeting(name: String, modifier: Modifier = Modifier) {
-  MainScreen()
+  val mapViewModel: MapViewModel = viewModel()
+  MapScreen(mapViewModel)
 }
 
 @Preview(showBackground = true)
